commit2

+++ view_books.php
@@ -2,14 +2,32 @@
 <html>
     <head>
         <title>view books</title>
-
+        <meta charset="UTF-8">
+        <meta name="viewport" content="width=device-width, initial-scale=1.0">
+        <meta http-equiv="X-UA-Compatible" content="ie=edge">
+        
+        <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
     </head>
     <body>
         <div class="head">
-        <h1>Books available in Library</h1>
+            <h1>Books available in Library</h1>
         </div>
+        <nav>
+            <form action="view_books.php" method="post">
+                <input type="text" name = 'search' id="search" placeholder="What are you looking for ?">
+                <input type="submit" name="submit" value="Search">
+            </form>
+        </nav>
+        <br>
+        <br>
+        <br>
+        <br>
+        <br>
+        <br>
+
+
         
-            <?php
+        <?php
             session_start();
             if (!isset($_SESSION['regno'])) {
                 // Redirect to login page
@@ -17,25 +35,50 @@
                 exit();
             }
             include("DB_connect.php");
-            $sql = "SELECT * FROM books ORDER BY 'bookname' DESC";
-            $result = $conn->query($sql);
-            if ($result->num_rows > 0) {
-            // output data of each row
-            echo "<table>";
-            echo"<tr>
-                <th>Book Id</th>
-                <th>Book Name</th>
-                <th>Author</th>
-                <th>Edition</th>
-            </tr>";
-                while($row = $result->fetch_assoc()) {
+            if(isset($_POST['submit'])){
+                $search = mysqli_real_escape_string($conn, $_POST['search']);
+                $sql = "SELECT * FROM books WHERE bname LIKE '%$search%'";
+                $result = $conn->query($sql);
+                if ($result->num_rows > 0) {
+                // output data of each row
+                echo "<table>";
+                echo"<tr>
+                    <th>Book Id</th>
+                    <th>Book Name</th>
+                    <th>Author</th>
+                    <th>Edition</th>
+                </tr>";
+                    while($row = $result->fetch_assoc()) {
 
-                    echo "<tr><td>" . $row["bid"]. "</td><td>" . $row["bname"] . "</td><td>". $row["author"]. "</td><td>". $row["edition"]. "</td></tr>";
+                        echo "<tr><td>" . $row["bid"]. "</td><td>" . $row["bname"] . "</td><td>". $row["author"]. "</td><td>". $row["edition"]. "</td></tr>";
+                    }
+                    echo "</table>";
+                } else { echo "<script>if(window.confirm('Book you searched is Not available in the Library.....')){
+                    window.location.href = 'view_books.php'}else{}</script>"; }
+            }else{
+                $sql = "SELECT * FROM books ORDER BY 'bookname' DESC";
+                $result = $conn->query($sql);
+                if ($result->num_rows > 0) {
+                // output data of each row
+                echo "<table>";
+                echo"<tr>
+                    <th>Book Id</th>
+                    <th>Book Name</th>
+                    <th>Author</th>
+                    <th>Edition</th>
+                    <th>E-Book</th>
+                </tr>";
+                    while($row = $result->fetch_assoc()) {
+
+                        echo "<tr><td>" . $row["bid"]. "</td><td>" . $row["bname"] . "</td><td>". $row["author"]. "</td><td>". $row["edition"]. "</td><td><a href=" . $row['url'] . ">E-Book</a></td></tr>";
+                    }
+                    echo "</table>";
+                } else { echo "<script>if(window.confirm('Books are Not available in the Library.....')){
+                    window.location.href = 'Home.html'}else{}</script>"; }
                 }
-                echo "</table>";
-            } else { echo "<script>if(window.confirm('Books are Not available in the Library.....')){
-                window.location.href = 'Home.html'}else{}</script>"; }
-            ?>
+
+            
+        ?>
         
         
     </body>
@@ -53,6 +96,38 @@
             background: rgb(241, 241, 241);
             background-size: cover;
         }
+
+        nav{
+            
+            float: right;
+            word-spacing: 20px;
+            padding: 20px;
+        }
+
+        nav input[type="text"]{
+            width: 200px;
+            height: 30px;
+            border: 1px solid black;
+            border-radius: 5px;
+            outline: none;
+            padding: 5px;
+        }
+
+        nav input[type="text"]:focus{
+            border: 1px solid #588c7e;
+        }
+
+        nav input[type="submit"]{
+            width: 100px;
+            height: 30px;
+            border: 1px solid black;
+            border-radius: 5px;
+            outline: none;
+            padding: 5px;
+            background-color: #588c7e;
+            color: white;
+        }
+        
         table {
             margin: auto;
             font-size: medium;
@@ -71,5 +146,7 @@
             color: white;
         }
         tr:nth-child(even) {background-color: #ecf8f5}
+
+        
     </style>
 </html>
